requirements.txt
# Bibliotecas necessárias para o Gerador de Excel com Tkinter
tk
Pillow
openpyxl

